backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.40.0
pyodbc==5.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.9.0
simple-salesforce>=1.12.0
